refactor: extract tui file-backed tests
